Done builder

+++ src/index.ts
@@ -2,7 +2,7 @@ import * as http from "http";
 import * as url from "url";
 import * as fs from "fs-extra";
 
-import { PORT, TEMP_DIR } from "./config";
+import { PORT, TEMP_DIR, EMPTY_TEMP_DIR_BEFORE_BUILD } from "./config";
 import handleBuildWithGithubRepo from "./pages/build-with-github-repo";
 import handleBuildWithZip from "./pages/build-with-zip";
 
@@ -20,9 +20,7 @@ function startServer() {
     try {
       let requestUrl = url.parse(request.url);
       let params = new url.URLSearchParams(requestUrl.query);
-  
-      console.log("pathname= " + requestUrl.pathname);
-      console.log("params= " + params);
+      console.timeLog("processing request: " + request.url);
       switch (requestUrl.pathname) {
         case "/build-with-github-repo":
           handleBuildWithGithubRepo(request, response, params);
@@ -37,10 +35,14 @@ function startServer() {
       console.error(error);
     }
   });
-  console.log("[" + new Date().toISOString() + "] listening port at: " + PORT);
+  console.timeLog("listening port at: " + PORT, true);
   server.listen(PORT);
 }
 
+console.timeLog = (msg: string) => console.log("[" + new Date().toLocaleString() + "] " + msg);
+
 fs.ensureDirSync(TEMP_DIR);
-fs.emptyDirSync(TEMP_DIR);
+if (EMPTY_TEMP_DIR_BEFORE_BUILD) {
+  fs.emptyDirSync(TEMP_DIR);
+}
 startServer();
